Passage triangulation 2D en doubles
git-svn-id: http://rdinnovation.onf.fr/svn/computree@362 0d0f61bb-bbfd-4940-b5ed-fa75979e1372

+++ pluginshared/ct_triangulation/ct_voronoinodet.cpp
@@ -1,23 +1,23 @@
 #include "ct_voronoinodet.h"
 
-CT_VoronoiNodeT::CT_VoronoiNodeT(float x, float y, float z)
+CT_VoronoiNodeT::CT_VoronoiNodeT(double x, double y, double z)
 {
     _x = x;
     _y = y;
     _z = z;
 }
 
-float CT_VoronoiNodeT::getX() const
+double CT_VoronoiNodeT::getX() const
 {
     return _x;
 }
 
-float CT_VoronoiNodeT::getY() const
+double CT_VoronoiNodeT::getY() const
 {
     return _y;
 }
 
-float CT_VoronoiNodeT::getZ() const
+double CT_VoronoiNodeT::getZ() const
 {
     return _z;
 }
